window: Update slider position on change-value
Currently, the slider position is updated in response
to the GstPlay player position update. Therefore, when
the user seeks, the slider widget is updated only once
playback is resumed. This results in a non-smooth UI/UX.

This commit add slider position update as part of the 'change-value'
event handler. This means the position and label of the slider
will be updated as soon as the user seeks, and even while the user
is seeking.