Check all supported CODECs for hardware-acceleration
The v4l2codecs plugin supports a bunch of CODECs, add them all.
At the time of writing this, HEVC acceleration is a work-in-progress
and not merged.

However, the plugin is in good shape, so it will
be merged as 'v4l2slh265dec' as some point, so it makes sense
to track it.

See: https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/1079