window: Add ref_uri
This is the name we want to refer to the stream. For local files it
makes no difference. For https:// streams that pass through a URL
processor the ref_uri is the initial short name and the uri is the
actual playable backend uri (which e.g. for video sites changes with
every play).