sink: Add support for DMABuf import and graphics offload
By implementing support for `GdkDmabufTextureBuilder` and
`GstVideoInfoDmaDrm`. This allows zero-copy video playback on Wayland
when paired with hardware video decoding.

Mostly a straigh forward copy from `GtkGstSink`.

Note that Gstreamer 1.23.1/1.24 is required.

We bump the required GTK version due to the change in the UI file, which
is hard to make conditional.